feature: 连接管理器，管理连接对生命周期
